requirements.txt
eng_to_ipa==0.0.2
ipapy==0.0.9.0
nltk==3.5
absl-py==0.9.0
astor==0.8.1
audioread==2.1.5
certifi==2020.4.5.1
cffi==1.14.0
cycler==0.10.0
decorator==4.4.2
falcon==1.2.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
inflect==0.2.5
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
librosa==0.6.1
llvmlite==0.32.0
lws==1.2.6
Markdown==3.2.1
matplotlib==2.0.2
mock==4.0.2
numba==0.39.0
numpy==1.14.0
opt-einsum==3.2.1
protobuf==3.11.3
PyAudio==0.2.11
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2019.3
PyYAML==5.3.1
resampy==0.2.2
scikit-learn==0.22.2.post1
scipy==1.0.0
six==1.14.0
sounddevice==0.3.10
SoundFile==0.10.3.post1
speedtest-cli==2.1.2
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
tqdm==4.11.2
Unidecode==0.4.20
Werkzeug==1.0.1
wrapt==1.12.1
